move angle and time
