--- packages/gui/src/components/inputs/TimeInput.tsx
@@ -1,24 +1,0 @@
-import { Time, TIME_UNITS } from '../../types/css'
-import { EditorPropsWithLabel } from '../../types/editor'
-import { DimensionInput } from './Dimension'
-
-export const TimeInput = (props: EditorPropsWithLabel<Time>) => {
-  return (
-    <DimensionInput
-      units={TIME_UNITS}
-      steps={timeSteps}
-      conversions={timeConversions}
-      {...props}
-    />
-  )
-}
-
-const timeConversions = {
-  ms: 1000,
-  s: 1,
-}
-
-export const timeSteps = {
-  ms: 25,
-  s: 0.025,
-}
